# src/DbBackupManager.Web/OpenApi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DbBackupManager.Web | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/backup-overview": {
      "get": {
        "tags": [
          "BackupMonitoring"
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupOverviewResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminCookie": [ ]
          }
        ]
      }
    },
    "/api/v1/backup-tasks": {
      "get": {
        "tags": [
          "BackupMonitoring"
        ],
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "maximum": 10000,
              "minimum": 0,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "maxLength": 200,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "maxLength": 30,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "FromUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "UntilUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "OldestFirst",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupTaskPageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminCookie": [ ]
          }
        ]
      }
    },
    "/api/v1/backup-files": {
      "get": {
        "tags": [
          "BackupMonitoring"
        ],
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "maximum": 10000,
              "minimum": 0,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "maxLength": 200,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "maxLength": 30,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "FromUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "UntilUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "OldestFirst",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupFilePageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminCookie": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackupFilePageResponse": {
        "required": [
          "items",
          "page",
          "totalCount",
          "pageSize",
          "hasMore"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BackupFileSummaryResponse"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
      "BackupFileSummaryResponse": {
        "required": [
          "taskId",
          "databaseId",
          "databaseName",
          "instanceName",
          "policyName",
          "sqlPath",
          "lengthBytes",
          "verifiedAtUtc",
          "retentionDays",
          "fileId",
          "location",
          "protocol",
          "targetDisplayName",
          "status",
          "deletedAtUtc",
          "errorCode",
          "path"
        ],
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "format": "uuid"
          },
          "databaseId": {
            "type": "string",
            "format": "uuid"
          },
          "databaseName": {
            "type": "string"
          },
          "instanceName": {
            "type": "string"
          },
          "policyName": {
            "type": "string"
          },
          "sqlPath": {
            "type": "string"
          },
          "lengthBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "verifiedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "retentionDays": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "fileId": {
            "type": "string",
            "format": "uuid"
          },
          "location": {
            "type": "string"
          },
          "protocol": {
            "type": "string"
          },
          "targetDisplayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          },
          "deletedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "errorCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "path": {
            "type": "string"
          }
        }
      },
      "BackupOverviewResponse": {
        "required": [
          "generatedAtUtc",
          "recentFromUtc",
          "managedDatabases",
          "withoutEnabledPolicy",
          "neverVerified",
          "withVerifiedFile",
          "pendingTasks",
          "runningTasks",
          "needsAttentionTasks",
          "recentFailedTasks",
          "worker",
          "recentTasks",
          "recentFailures",
          "recentFiles"
        ],
        "type": "object",
        "properties": {
          "generatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "recentFromUtc": {
            "type": "string",
            "format": "date-time"
          },
          "managedDatabases": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "withoutEnabledPolicy": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "neverVerified": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "withVerifiedFile": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pendingTasks": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "runningTasks": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "needsAttentionTasks": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "recentFailedTasks": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "worker": {
            "$ref": "#/components/schemas/WorkerPresenceResponse"
          },
          "recentTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BackupTaskSummaryResponse"
            }
          },
          "recentFailures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BackupTaskSummaryResponse"
            }
          },
          "recentFiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BackupFileSummaryResponse"
            }
          }
        }
      },
...
      "BackupTaskPageResponse": {
        "required": [
          "items",
          "page",
          "totalCount",
          "pageSize",
          "hasMore"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BackupTaskSummaryResponse"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
      "BackupTaskSummaryResponse": {
        "required": [
          "id",
          "policyName",
          "databaseName",
          "status",
          "stage",
          "createdAtUtc",
          "completedAtUtc",
          "errorCode",
          "cancellationRequested",
          "fileLength"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "policyName": {
            "type": "string"
          },
          "databaseName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "stage": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "completedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "errorCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "cancellationRequested": {
            "type": "boolean"
          },
          "fileLength": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "WorkerPresenceResponse": {
        "required": [
          "status",
          "lastSeenAtUtc",
          "offlineAfterSeconds"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "lastSeenAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "offlineAfterSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      }
    },
    "securitySchemes": {
      "AdminCookie": {
        "type": "apiKey",
        "description": "管理员登录后由同源 Web 主机签发的安全 Cookie。",
        "name": "DbBackupManager.Admin",
        "in": "cookie"
      }
    }
  },
...
}